requirements.txt
streamlit
pandas
openpyxl # Excelファイルの読み書きに必要
xlsxwriter # Excelファイルへの書き込み(ハイライト付き保存)に必
google-generativeai
python-amazon-sp-api
python-dotenv